apps(requests): move org related code to the lib
